Slight refactoring
