+++ pkgbuild.go
@@ -1,0 +1,5 @@
+package main
+
+func fetchPkgBuild(pkg string) error {
+	return nil
+}
